I uploaded my files

+++ contents/logout.php
@@ -1,9 +1,11 @@
 <?php
     include './dbconnect.php';
 
-    $username = $_SESSION['username'];
-    if($_SESSION['createtable'] == 1){
-        $conn->query("DROP TABLE `$username`");
+    if(isset($_SESSION['username'])){
+        $username = $_SESSION['username'];
+        if($_SESSION['createtable'] == 1){
+            $conn->query("DROP TABLE `$username`");
+        }
     }
 
     session_destroy();
